move old cli+options to gui subdirectory and replace them with webview versions
