+++ bumps/__main__.py
@@ -5,7 +5,7 @@ Run "bumps --help" for details, or "python -m bumps -help" if
 bumps isn't on your path.
 """
 
-from bumps.webview.server.cli import main
-
 if __name__ == "__main__":
-    main()
+    from bumps.cli import main as new_main
+
+    new_main()
