+++ bumps/webview/server/__init__.py
@@ -1,1 +1,1 @@
-from .webserver import start_bumps_server
+# from .webserver import start_bumps_server # causes circular import from bumps.cli
